Add functionality of find movies by id

+++ src/main/java/com/julianolarte/netflix/repositories/MovieRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.repository.query.Param;
 
-public interface MovieRepository extends CrudRepository<Movie, Long> {
+public interface MovieRepository extends CrudRepository<Movie, Integer> {
 
     /**
      * @return Last 10 movies order by year Desc
